Issue #34,#27,#33: Initial Integration
Began work on putting together the SessionCollectionService, the HikeDataDirector and the HikeHardwareManager. Initial tests work in the backend code. Have problems with successful TI ST Connection.

+++ app/src/main/java/me/dotteam/dotprod/data/Coordinates.java
@@ -34,15 +34,15 @@ public class Coordinates {
         return retVal;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public float getAltitude() {
+    public double getAltitude() {
         return altitude;
     }
 }
